small re-organization of internal ecephys tests
